More AV1 parsing

+++ YUViewLib/src/parser/AV1/AV1OBU.h
@@ -50,11 +50,10 @@ public:
   ParserAV1OBU(QObject *parent = nullptr);
   ~ParserAV1OBU() {}
 
-  std::pair<unsigned, std::string> parseAndAddOBU(int         obuID,
-                                                  ByteVector &data,
-                                                  TreeItem *  parent,
-                                                  pairUint64  obuStartEndPosFile = pairUint64(-1,
-                                                                                             -1));
+  std::pair<size_t, std::string> parseAndAddOBU(int         obuID,
+                                                ByteVector &data,
+                                                TreeItem *  parent,
+                                                pairUint64 obuStartEndPosFile = pairUint64(-1, -1));
 
   // So far, we only parse AV1 Obu files from the AVFormat parser so we don't need this (yet).
   // When parsing of raw OBU files is added, we will need this.
